requirements.txt
fastapi
uvicorn[standard]
python-multipart
streamlit
requests
python-dotenv

langchain
langgraph
langchain-openai
langchain-community
langchain-chroma
langchain-text-splitters

chromadb
pypdf
tiktoken

pytest